Cleanup old unused tests
